+++ CHANGELOG.md
@@ -10,6 +10,7 @@ CHANGELOG for FlatCAM Evo beta
 22.05.2022
 
 - in Gerber Editor upgraded the PadAdd GUI
+- in Gerber Editor upgraded the PadArray GUI
 
 21.05.2022